Update patch set 1
Patch Set 1: No score; No score

To be honest, I'm not sure I understand why this patch is so big, and why it changes so many things. The only thing you really want from this patch is the ability to provide the charge function from the board file. None of the other changes to ds2784 or the w1 code seems necessary.

Patch-set: 1
Label: Code-Review=0
Label: Verified=0