Update patch set 1
Patch Set 1: No score; No score

The w1 ds2784 driver previously created the platform_device for the battery driver.  Because of that, we couldn't pass board-specific information to ds2784_battery, and it was hard-coded (GPIOs) in the file itself.  The reorg of w1/ds2784 allows the ds2784_battery code to instantiate the latter, allowing us to put a platform_device struct in the board file.

Patch-set: 1
Label: Code-Review=0
Label: Verified=0